Rename: 폴더명 model -> dto 로 변경
