+++ src/test/java/com/zerobase/BankSSun/service/UserServiceTest.java
@@ -3,7 +3,7 @@ package com.zerobase.BankSSun.service;
 import static org.junit.jupiter.api.Assertions.*;
 
 import com.zerobase.BankSSun.domain.entity.UserEntity;
-import com.zerobase.BankSSun.model.Auth.SignUp;
+import com.zerobase.BankSSun.dto.Auth.SignUp;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
